requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
astroid==2.6.2
attrs==17.4.0
Automat==0.6.0
blinker==1.4
cairocffi==1.0.0
CairoSVG==2.5.2
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.5
django-braces==1.14.0
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-widget-tweaks==1.4.8
docopt==0.6.2
html5lib==1.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
isort==5.9.2
jeepney==0.6.0
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.6.0
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==2.0.3
netifaces==0.10.4
num2words==0.5.10
oauthlib==2.0.6
pexpect==4.2.1
Pillow==8.3.1
pkg-resources==0.0.0
ply==3.11
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
PyJWT==1.5.3
pylint==2.9.3
pymemcache==3.5.0
pyOpenSSL==17.5.0
pyphen==0.11.0
pypng==0.0.20
PyQRCode==1.2.1
pyserial==3.4
python-debian==0.1.32
python-decouple==3.4
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
sqlparse==0.4.1
ssh-import-id==5.7
tinycss2==1.1.0
toml==0.10.2
Twisted==17.9.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.22
WeasyPrint==52.5
webencodings==0.5.1
wrapt==1.12.1
zope.interface==4.3.2
